+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 ### Added
 - Added the PHP `password_hash()` function as a mandatory requirement.
+- Added the PHP Zip extension as a mandatory requirement.
 
 ## 1.0.8 - 2017-02-06